% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subanalysis_flags.R
\name{add_subanalysis_flags}
\alias{add_subanalysis_flags}
\title{Flag patients for subanalysis, stratified on exposure to approved drugs for target disease
Last updated: 2021-02-10}
\usage{
add_subanalysis_flags(indication_drug_exposed, covariates, ...)
}
\arguments{
\item{indication_drug_exposed}{A tibble with columns, c('person_id','rxcui_ingr','rxcui_ingr_name','group')}

\item{covariates}{A tibble with columns, c('person_id','dob','gender_concept_id','race_concept_id','start_date','first_drug_exposure','final_end_date','age')}
}
\description{
Flag patients for subanalysis, stratified on exposure to approved drugs for target disease
Last updated: 2021-02-10
}
