% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_clinical_data.R
\name{analyze_clinical_data}
\alias{analyze_clinical_data}
\title{Processes clinical data and outputs table summarizing treatment effect and number of overlapping patients in raw data
Last updated: 2021-02-17}
\usage{
analyze_clinical_data(
  drug,
  phenotype,
  biomarker,
  indication_drugs,
  concept_id_exclusions,
  r_covariates,
  r_drugs,
  r_biomarkers
)
}
\arguments{
\item{drug}{A string, e.g., "simvastatin"}

\item{phenotype}{A string, e.g., "Hyperlipidemia"}

\item{biomarker}{A string, e.g., "LDL Cholesterol"}

\item{indication_drugs}{A tibble with columns c('rxcui_in', 'drug_desc', 'indication'), e.g.,  systematicRepurposing::lipid_drugs_20201217141539}

\item{concept_id_exclusions}{A vector, e.g., c(2212451)}

\item{r_covariates}{A tibble with columns c('person_id','dob','gender_concept_id','race_concept_id','drug_concept_id','first_drug_exposure','last_drug_exposure')}

\item{r_drugs}{A tibble with columns c('person_id','start_date','end_date','drug_exposure_start_date','drug_concept_id')}

\item{r_biomarkers}{A tibble with columns c('person_id','start_date','measurement_date','end_date','measurement_concept_id','value_as_number')}
}
\description{
Processes clinical data and outputs table summarizing treatment effect and number of overlapping patients in raw data
Last updated: 2021-02-17
}
