% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_process_clinical_data.R
\name{clean_process_clinical_data}
\alias{clean_process_clinical_data}
\title{Cleans and processes clinical data extracted from EHR formatted in OHDSI/OMOP CDM. Output is a tibble with each row representing one patient.
This function was created to create tibble with only patients not exposed to known approve drugs for the phenotype of interest, besides the drug being tested,
ie, if "simvastatin" is being tested for its effect on treating hyperlipidemia, "simvastatin" would be removed from the indication drug list, but
other statins would still remain.
Last updated: 2021-03-13}
\usage{
clean_process_clinical_data(
  drug,
  phenotype,
  biomarker,
  indication_drugs,
  concept_id_exclusions,
  r_covariates,
  r_drugs,
  r_biomarkers
)
}
\arguments{
\item{drug}{A string, e.g., "simvastatin"}

\item{phenotype}{A string, e.g., "Hyperlipidemia"}

\item{biomarker}{A string, e.g., "LDL Cholesterol"}

\item{indication_drugs}{A tibble with columns c('rxcui_in', 'drug_desc', 'indication'), e.g.,  systematicRepurposing::lipid_drugs_20201217141539}

\item{concept_id_exclusions}{A vector, e.g., c(2212451)}

\item{r_covariates}{A tibble with columns c('person_id','dob','gender_concept_id','race_concept_id','drug_concept_id','first_drug_exposure','last_drug_exposure')}

\item{r_drugs}{A tibble with columns c('person_id','start_date','end_date','drug_exposure_start_date','drug_concept_id')}

\item{r_biomarkers}{A tibble with columns c('person_id','start_date','measurement_date','end_date','measurement_concept_id','value_as_number')}
}
\description{
Cleans and processes clinical data extracted from EHR formatted in OHDSI/OMOP CDM. Output is a tibble with each row representing one patient.
This function was created to create tibble with only patients not exposed to known approve drugs for the phenotype of interest, besides the drug being tested,
ie, if "simvastatin" is being tested for its effect on treating hyperlipidemia, "simvastatin" would be removed from the indication drug list, but
other statins would still remain.
Last updated: 2021-03-13
}
