% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_observation_periods.R
\name{define_observation_periods}
\alias{define_observation_periods}
\title{Define baseline and treatment periods
Last updated: 2021-02-10}
\usage{
define_observation_periods(.data, ...)
}
\arguments{
\item{.data}{A tibble}
}
\description{
Baseline period = "start_date" to "first_{drug}\emph{exposure". Treatment period =
("first}{drug}\emph{exposure" + 30 days) to "last}{drug}\emph{exposure". If treatment period > 12
months, then treatment period = ("first}{drug}_exposure" + 30 days) to "end_date"
}
