% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clinical_data.R
\name{extract_clinical_data}
\alias{extract_clinical_data}
\title{Function to extract clinical data from All of Us dataset
Last updated: 2021-03-10}
\usage{
extract_clinical_data(
  drug_concept_id,
  biomarker_concept_id,
  table_name,
  table_type,
  indication_drug_concept_ids
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{table_name}{A string}

\item{table_type}{A string, either c("covariates", "drugs", "biomarkers")}

\item{indication_drug_concept_ids}{A string, with concept_ids for indication drugs, eg,
"({drug_concept_id},1551803,19011061,1351115,1558242,19004048,19051463,19018419,1510813,
1501617,1592085,1545958,1549686,46275447,1551860,19082770,1586226,19095309,46274162,957797,
1517824,46287466,19022956,40165636,1518148,19026600,1539403,1526475)"}
}
\description{
Function to extract clinical data from All of Us dataset
Last updated: 2021-03-10
}
