% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observation_period_drug_exposures.R
\name{observation_period_drug_exposures}
\alias{observation_period_drug_exposures}
\title{Get baseline and treatment period drug exposures
Last updated: 2021-02-10}
\usage{
observation_period_drug_exposures(df_drug_obsperiod, drugs, ...)
}
\arguments{
\item{df_drug_obsperiod}{A tibble with columns, c('person_id','start_date','first_drug_exposure','final_end_date')}

\item{drugs}{A tibble with columns, c('person_id','rxcui_ingr','rxcui_ingr_name','drug_exposure_start_date')}
}
\description{
Get baseline and treatment period drug exposures
Last updated: 2021-02-10
}
