% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_covariates.R
\name{process_covariates}
\alias{process_covariates}
\title{Process covariates data: keep only adult patients, calculate age at start of observation period, and keep only European ancestry patients
Last updated: 2021-02-10}
\usage{
process_covariates(covariates, df_drug_obsperiod, european = 1, ...)
}
\arguments{
\item{covariates}{A tibble with columns, c('person_id','dob','gender_concept_id','race_concept_id')}

\item{df_drug_obsperiod}{A tibble with columns, c('person_id','start_date','first_drug_exposure','final_end_date')}

\item{european}{A binary value with 1 = only use european ancestry, and 0 = use entire cohort.}
}
\description{
Process covariates data: keep only adult patients, calculate age at start of observation period, and keep only European ancestry patients
Last updated: 2021-02-10
}
