% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results_table.R
\name{show_results_table}
\alias{show_results_table}
\title{Show results table
Last updated: 2021-02-14}
\usage{
show_results_table(data_table)
}
\arguments{
\item{data_table}{A tibble}
}
\description{
Show results table
Last updated: 2021-02-14
}
