% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_clinical_data.R
\name{extract_clinical_data}
\alias{extract_clinical_data}
\title{Function to extract clinical data from All of Us dataset
Last updated: 2021-03-10}
\usage{
extract_clinical_data(
  drug_concept_id,
  biomarker_concept_id,
  table_name,
  table_type,
  conn,
  ...
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{table_name}{A string}

\item{table_type}{A string, either c("covariates", "drugs", "biomarkers")}

\item{conn}{A variable}
}
\description{
Function to extract clinical data from All of Us dataset
Last updated: 2021-03-10
}
